performance: use double instead of ints, saves a casting until we support discrete bins

+++ packages/vaex-core/src/hash_object.cpp
@@ -191,7 +191,7 @@ public:
         }
         return result;
     }
-    py::array_t<int64_t> map_ordinal_with_mask(py::buffer object_array, py::array_t<bool>& masks) {
+    py::array_t<double> map_ordinal_with_mask(py::buffer object_array, py::array_t<bool>& masks) {
         auto m = masks.template unchecked<1>();
         py::buffer_info info = object_array.request();
         if(info.ndim != 1) {
@@ -199,7 +199,7 @@ public:
         }
         // TODO: check dtype/format
         int64_t size = info.shape[0];
-        py::array_t<int64_t> result(size);
+        py::array_t<double> result(size);
         auto output = result.template mutable_unchecked<1>();
         assert(m.size() == size);
         PyObject** array = (PyObject**)info.ptr;
